refactor(mpc): Move storage types into same file, add missing license header
