+++ packages/mpc/src/actions/keygen.ts
@@ -9,7 +9,8 @@ import {
 import _sodium, { base64_variants } from "libsodium-wrappers-sumo";
 import { MpcError, MpcErrorCode } from "../error";
 import type { HttpClient } from "../transport/httpClient";
-import type { KeygenConversation, PairingData } from "../types";
+import type { KeygenConversation } from "../types";
+import type { PairingData } from "../storage/types";
 import * as utils from "../utils";
 
 type KeygenResult = {
