+++ packages/mpc/src/constants.ts
@@ -1,3 +1,5 @@
+// Copyright (c) Silence Laboratories Pte. Ltd.
+// This software is licensed under the Silence Laboratories License Agreement.
 
 export enum StoragePlatform {
 	Browser = "browser",
