+++ packages/mpc/src/index.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Silence Laboratories Pte. Ltd.
 // This software is licensed under the Silence Laboratories License Agreement.
 
-import type { AccountData, StorageData } from "./types";
+import type { AccountData, StorageData } from "./storage/types";
 import type { IStorage } from "./storage/types";
 import { StoragePlatform, WalletId,  } from "./constants";
 import { MpcAuthenticator } from "./domain/authenticator";
