+++ packages/mpc/src/types.ts
@@ -63,50 +63,5 @@ export interface BackupConversation {
 	walletId: string;
 }
 
-export interface PairingData {
-	pairingId: string;
-	webEncPublicKey: string;
-	webEncPrivateKey: string;
-	webSignPublicKey: string;
-	webSignPrivateKey: string;
-	token: string;
-	tokenExpiration: number;
-	appPublicKey: string;
-	deviceName: string;
-}
-
-export interface DistributedKey {
-	accountId: number;
-	publicKey: string;
-	keyShareData: IP1KeyShare;
-}
-
-export type V0StorageData = {
-	pairingData: PairingData;
-	newPairingState?: {
-		pairingData: PairingData | null;
-		distributedKey: DistributedKey | null;
-	};
-};
-
-export interface StorageData {
-	version?: number;
-	pairingData: PairingData | null;
-	distributedKey: DistributedKey | null;
-	walletAccount?: AccountData;
-	eoa: string | null;
-	passwordReady?: boolean;
-}
-export interface PairingSessionData {
-	token: string;
-	appPublicKey: string;
-	deviceName: string;
-	tokenExpiration: number;
-	backupData?: string | undefined;
-}
-
 export type DeviceOS = "ios" | "android";
 
-export type AccountData = {
-	address: string;
-};
